ctree/requirements.txt
uwsgi
django
djangorestframework
psycopg2-binary
